requirements.txt
Flask==2.2.2
gunicorn==20.1.0
python-dotenv===1.0.1
b2sdk===2.7.0